[searx-search] move module under utilities
- Make the module loadable the same way as other utilities
