+++ langchain/__init__.py
@@ -33,6 +33,7 @@ from langchain.prompts import (
 from langchain.serpapi import SerpAPIChain, SerpAPIWrapper
 from langchain.sql_database import SQLDatabase
 from langchain.utilities.google_search import GoogleSearchAPIWrapper
+from langchain.utilities.searx_search import SearxSearchWrapper
 from langchain.utilities.wolfram_alpha import WolframAlphaAPIWrapper
 from langchain.vectorstores import FAISS, ElasticVectorSearch
 
@@ -48,6 +49,7 @@ __all__ = [
     "SelfAskWithSearchChain",
     "SerpAPIWrapper",
     "SerpAPIChain",
+    "SearxSearchWrapper",
     "GoogleSearchAPIWrapper",
     "WolframAlphaAPIWrapper",
     "Anthropic",
