+++ langchain/utilities/__init__.py
@@ -1,11 +1,11 @@
 """General utilities."""
 from langchain.python import PythonREPL
 from langchain.requests import RequestsWrapper
-from langchain.searx_search import SearxSearchWrapper
 from langchain.serpapi import SerpAPIWrapper
 from langchain.utilities.bash import BashProcess
 from langchain.utilities.bing_search import BingSearchAPIWrapper
 from langchain.utilities.google_search import GoogleSearchAPIWrapper
+from langchain.utilities.searx_search import SearxSearchWrapper
 from langchain.utilities.wolfram_alpha import WolframAlphaAPIWrapper
 
 __all__ = [
